\name{blabla-package}
\alias{blabla-package}
\alias{blabla}
\docType{package}
\title{
\packageTitle{blabla}
}
\description{
\packageDescription{blabla}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{blabla}
\packageIndices{blabla}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{blabla}

Maintainer: \packageMaintainer{blabla}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the ~~
~~ R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
